apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "law-juridico-monitor-api"
version = "0.1.0"
description = "Backend do LAW Juridico Monitor"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.0",
  "apscheduler>=3.10.4",
  "email-validator>=2.2.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "openai>=1.68.0",
  "orjson>=3.10.0",
  "pandas>=2.3.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.7.0",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "rapidfuzz>=3.13.0",
  "redis>=5.2.1",
  "reportlab>=4.3.1",
  "sqlalchemy>=2.0.39",
  "structlog>=24.4.0",
  "tenacity>=9.1.2",
  "uvicorn[standard]>=0.34.0",
  "xlsxwriter>=3.2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
